+++ CHANGELOG.md
@@ -16,6 +16,7 @@
 - Recipe for Magento now supports locale configuration for `setup:static-content:deploy`. [#2040]
 - When symfony_env is set to dev, require-dev are not installed. [#2035]
 - Fixed exit status of rollback command when there are no releases to rollback to. [#2052]
+- When the second parameter $options passed to run() and runLocally(), use it to overwrite default env config. [#2165]
 
 
 ## v6.8.0
@@ -582,6 +583,7 @@
 - Fixed remove of shared dir on first deploy.
 
 
+[#2165]: https://github.com/deployphp/deployer/issues/2165
 [#2150]: https://github.com/deployphp/deployer/issues/2150
 [#2111]: https://github.com/deployphp/deployer/pull/2111
 [#2052]: https://github.com/deployphp/deployer/issues/2052